+++ lib/dag.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+# Core graph
 require_relative "dag/errors"
 require_relative "dag/edge"
 require_relative "dag/result"
@@ -8,10 +9,12 @@ require_relative "dag/failure"
 require_relative "dag/graph"
 require_relative "dag/graph/builder"
 require_relative "dag/graph/validator"
+
+# Workflow layer
 require_relative "dag/workflow/step"
 require_relative "dag/workflow/registry"
 require_relative "dag/workflow/definition"
-require_relative "dag/steps"
-require_relative "dag/run_callbacks"
-require_relative "dag/runner"
-require_relative "dag/loader"
+require_relative "dag/workflow/steps"
+require_relative "dag/workflow/run_callbacks"
+require_relative "dag/workflow/runner"
+require_relative "dag/workflow/loader"
